sway: firefox-nightly, add marks, bar nop, WL-1
This commit consists of miscellaneous changes to my sway config that
do not warrant individual commits.

Switches instances of firefox in the config to firefox-nightly for
native wayland web browsing.

Alters the mark bindsym to add instead of replace.

Adds nop bindsyms to the bar for the mouse wheel (button4 + button5).

Adds WL-1 in addition to X11-1 now that a workspace can be assigned
an output priority list. This is used for the 99:TEST workspace for
nested sessions.