zshrc: remove GDK_BACKEND=wayland
This removes `GDK_BACKEND=wayland` from the zshrc since it is was
causing issues with some applications. The only GTK application that
I use that does not default to using the wayland backend is
firefox-nightly so it makes more sense to just set the environment
variable for it instead (which is already done in test rest of my config
files anyway).