vim: reword listhars and remove some imaps
This removed trail from listchars in favor of space and also adds an
eol char.

The imaps for <ESC>[H and <ESC>[F have been removed because they were
causing substitutions to fail. I don't remember which terminal those
were Home and End for and it doesn't appear to be needed for my current
setup so just remove them.